+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 1.5.0 [In progress]
+1. [#165](https://github.com/influxdata/influxdb-client-go/pull/165) Allow overriding the http.Client for the http service.
 
 ## 1.4.0 [2020-07-17]
 ### Breaking changes